+++ graph/adapter-output-spring-data-neo4j-ogm/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/internal/Neo4jStatementRepository.kt
@@ -281,6 +281,8 @@ RETURN COUNT(edit) as cnt""")
 
     @Query("""$MATCH_STATEMENT WHERE (sub.`resource_id`=$subjectId OR sub.`literal_id`=$subjectId OR sub.`predicate_id`=$subjectId OR sub.`class_id`=$subjectId) AND rel.`predicate_id` = $predicateId AND (obj.`resource_id`=$objectId OR obj.`literal_id`=$objectId OR obj.`predicate_id`=$objectId OR obj.`class_id`=$objectId) $RETURN_STATEMENT LIMIT 1""")
     fun findBySubjectIdAndPredicateIdAndObjectId(subjectId: ThingId, predicateId: ThingId, objectId: ThingId): Optional<Neo4jStatement>
+
+    fun findAllByStatementIdIn(statementIds: Set<StatementId>): List<Neo4jStatement>
 }
 
 @QueryResult