+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/input/rest/PaperVerificationCommandController.kt
@@ -3,6 +3,7 @@ package eu.tib.orkg.prototype.statements.adapter.input.rest
 import eu.tib.orkg.prototype.statements.application.port.`in`.MarkAsVerifiedUseCase
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
 import org.springframework.http.HttpStatus
+import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PutMapping
@@ -16,12 +17,14 @@ class PaperVerificationCommandController(
     private val service: MarkAsVerifiedUseCase
 ) {
     @PutMapping("/{id}/metadata/verified")
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     fun markVerified(@PathVariable id: ThingId) {
         service.markAsVerified(id)
     }
 
     @DeleteMapping("/{id}/metadata/verified")
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     fun unmarkVerified(@PathVariable id: ThingId) {
         service.markAsUnverified(id)