+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/spi/StatementRepository.kt
@@ -20,6 +20,7 @@ import org.springframework.transaction.annotation.Transactional
 interface StatementRepository : EntityRepository<GeneralStatement, StatementId> {
     fun countStatementsAboutResource(id: ThingId): Long
     fun countStatementsAboutResources(resourceIds: Set<ThingId>): Map<ThingId, Long>
+    fun determineOwnership(statementIds: Set<StatementId>): Set<OwnershipInfo>
     // legacy methods:
     fun nextIdentity(): StatementId
     @Transactional
@@ -80,4 +81,9 @@ data class ResourceContributor(
     val createdBy: String, // FIXME: This should be ContributorId
     @JsonProperty("created_at")
     val createdAt: String // FIXME: This should be OffsetDateTime
+)
+
+data class OwnershipInfo(
+    val statementId: StatementId,
+    val owner: ContributorId,
 )